Stacks are templates now

+++ semester_2/hw_2/t_2/hw2.task2.cpp
@@ -8,7 +8,7 @@ int main()
     clog << "Console calculator: type an expression in postfix notation"
          << endl;
     StackMachine calculator;
-    int result = calculator.calculate(cin);
+    double result = calculator.calculate(cin);
     clog << "Result = ";
     cout << result << endl;
     return 0;
